Add Lintier - CLI to quickly scaffold an ESLint & Prettier setup in a TypeScript project.
Lintier is a CLI to automate setting up an ESLint & Prettier setup in a TypeScript project.  It provides options to install the airbnb styleguide, stylelint for linting styles, and lint-staged for linting pre-commit. Lintier installs the dependencies, creates the config files, and adds linting scripts to a project's package.json file.  Here's a gif:

![](https://github.com/josh-stillman/lintier/blob/main/lintier.gif?raw=true)

I also added a section for installation and setup for this bullet point.  It didn't seem to fit in with the other tools in the "tools" section.  

Thanks for your consideration!  Please let me know if you have any questions.
-Josh