Add `testcafe-community` & Edit `TestCafe`
Add the new testcafe-community plugin which is a fork from the original plugin with community gathered rules.  It extends the original plugin which only provided the global environment variables for TestCafe. The original has not seen an update since 2016 and is less prepared to support PRs for additional rules.  Therefore, I moved the original TestCafe down to the globals category since that is all it provides from the actual implementation `index.js`.

Thanks again for this awesome repo!